+++ lib/app.cfg
@@ -1,0 +1,5 @@
+Zachary Hall
+ZH@gmail.com
+12345
+-5.00
+/users/outla/Documents/WCCC/Semester 4/CPT-180/Assignment2/app.cfg
